Extract AddressCard and add skeleton ui to it
